Add PoC for Tests not inheriting from TestCase
This commit adds some modifications necessary to be able to write Tests
that do not extend TestCase but that implement a certain Interface.
Currently the Interface is located in a separate repo in my presonal
namespace but for a "correct" implementation that of course needs to be
moved to the phpunit-namespacce.

UnitTests now need to implement the interface UnitTest. There is a
concrete implementation available in the trait UnitTester that handles
all the necessary imlpementation details for the Interface but they can
of course also be implemented differntly. Especially the implementation
of the UnitTest::getTestMethods can be individualized so that different
methods can be used to identify a test method. The current
implementation uses methods whose lowercased names start with
`checkThat`.

The Interfaces and implementations are available at
https://github.com/heiglandreas/phpunitstub